requirements.txt
streamlit
google-generativeai
gspread
oauth2client
pandas
PyPDF2
python-dotenv
